page구조 단일화,tab버그 수정
