+++ src/Components/Common/menuList/MenuList.js
@@ -13,7 +13,7 @@ const DropdownMenu = styled.div`
   width: 200px;
   box-shadow: 0px 4px 8px rgba(0, 0, 0, 0.1);
 
-
+z-index: 999;
 
   background: #FFFFFF;
   border-radius: 12px;
